+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Make setter for `MetricReaderOptions.PeriodicExportingMetricReaderOptions`
+  property public.
+  ([#3184](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3184))
+
 ## 1.2.0-rc5
 
 Released 2022-Apr-12